MCPClient: check uploadDIP/ dir in Store DIP
If the user chooses "Do not upload DIP" at the "Upload DIP?" choice
point, then the DIP will not be moved to the uploadedDIPs/ directory but
will stay in its original location in uploadDIP/. However, the "Store
DIP" micro-service expects the DIP to be in the uploadedDIPs/ directory. This
changes the storeAIP.py micro-services so that if the aip_path does not exist,
the path is modified so that the first uploadedDIPs/ directory in the
AIP path is replaced by uploadDIP. Not the most elegant solution, but it works.
Fixes #729.