MCPServer: unify distinct but identical choices
Assume that "decision point" means "a situation where one has options" and
"decision" means "the course of action that is ultimately pursued".
Archivematica allows for the creation of distinct decision points and decisions
(i.e., model instances or db rows) which are in fact semantically identical.
This commit allows one decision (point) to be designated as canonical.
The configuration of a canonical decision (point) can now be made to apply to
all of its counterparts.

Introduces the utils.py::choice_unifier dict that maps decision
point and decision UUIDs to their designated canonical equivalents. This
allows a processing MCP XML configuration file to reference UUIDs for
canonical decision points/decisions and have the equivalent
points/decisions be configured also.