Disable pool and connection recycling
In this commit we make the following changes:

- Set up dashboard to use the `django.db.backends.mysql` backend again (I think
that the change to `django_mysqlpool` was only an accident during the config
cleanup in `qa/1.x`) and explicitly give `CONN_MAX_AGE` a value of 0 because
that's needed when using gevent (greenlets), otherwise database connections
will accumulate.

- `MCPServer` and `MCPClient` to start using `django.db.backends.mysql` again.
`django_mysqlpool` is abandoned and incompatible with Py3. We make sure threads
close database connections before they terminate. Not recycling connections
brings some extra overhead (opening connections is expensive) but we believe
the majority of scalability issues origin in the total number of connections
concurrently open. In #553 we discoverd that `django_mysqlpool` wasn't really
helping to fix the latter. This is an issue that needs to be addressed
separately, with or without a pool, i.e. contention before the thread opens a
connection.