Delete database connection pool
This is a sequel of #505 which was originally opened to replace the
`django_mysqlpool` backend with the one provided by Django:
`django.db.backends.mysql`. I believe it was based on the erroneous assumption
that `django.db.backends.mysql` offered a database connection pool. What it
does is to provide persistent connections which are not a solution for web
servers using greenlets (Dashboard) or threaded applications like MCPServer or
MCPClient.

I think that deleting `django_mysqlpool` is still a good thing because it
doesn't support Py3 and it's pretty much abandoned. In this PR,
MCP{Server,Client} start using `django.db.backends.mysql` again. The connection
pool allowed us to recycle connections (opening connections is expensive) but it
didn't solve what I think is the biggest scalability issue: the total number of
connections opened simultaneously (see #553 for context). That's an issue that
needs to be addressed separately, e.g. implementing contention before the thread
opens a connection.

In this PR we also set up the Dashboard to use the `django.db.backends.mysql`
backend again (I think that the change to `django_mysqlpool` was only an
accident during the config cleanup in `qa/1.x`) and explicitly give
`CONN_MAX_AGE` a value of 0 because that's needed when using gevent
(greenlets), otherwise database connections will accumulate.