Backward-compatible logging
AM 1.7 brought a new simpler logging configuration with less responsibilities -
log writing to files, debugging log file or automatic file rotation are gone
now. The new default is to send the events to the standard streams which is
more convenient when you're running Archivematica in a cluster.

In this pull request I add the ability to change the new logging configuration
defaults via a JSON configuration file. The examples included mimic the old
behaviour - this is important for deb/pkg/ansible installations that still
prefer to have the old defaults in place.

- `/etc/archivematica/dashboard.logging.json` (see
[install/README.md](https://github.com/artefactual/archivematica/blob/dev/backward-compatible-logging/src/dashboard/install/README.md))
- `/etc/archivematica/clientConfig.logging.json` (see
[install/README.md](https://github.com/artefactual/archivematica/blob/dev/backward-compatible-logging/src/MCPClient/install/README.md))
- `/etc/archivematica/serverConfig.logging.json` (see
[install/README.md](https://github.com/artefactual/archivematica/blob/dev/backward-compatible-logging/src/MCPServer/install/README.md))

This closes #776.