MCPClient: disable capture of output for compressAIP client script
This adds a capture_output parameter to the functions in
executeOrRunSubprocess.py. This is set to True by default so existing
behaviour is maintained. However the compressAIP script sets capture_output
to False, thereby disabling the capture of stdout and stderr for this
specific task.

Includes unit test to demonstrate expected functionality.

This fixes JiscRDSS issue #39.