dashboard: make create_super_user idempotent in installer steps
If a user already exists with the same name as the given user then the
create_super_user function in steps.py would fail because the user already
exists. This change makes the method idempotent, so that any existing user
will be ignored and not cause an error.

This fixes Jisc issue #45.