Add common abilities with choice dependencies
New abilities.py module in archivematicaCommon defines abilities of an
Archivematica instance as a whole. An ability can have dependencies. For
example, the Backlog feature is dependent on the Search ability. If the Search
(i.e., Elasticsearch indexing) ability is disabled, then the Backlog tab should
be unavailable; in addition, the choice to send a transfer to backlog (either
via pre-configuration or via real-time choice in the GUI) should not be
available either.

Some minor formatting and i18n changes were also made.