Enhance appconfig to allow multiple equivalent options
The `appconfig.py::Config` class accepts an `attrs` dict argument whose values
can now contain lists of dicts as opposed to just single dicts as previsouly.
This allows one to specify multiple configuration options as equivalent and
reduce them to a single option. Using this functionality, the
`disable_search_indexing` and `search_enabled` config options are declared to
be equivalent, the one being the inverse of the other. The resulting settings
attribute is accessed as `settings.SEARCH_ENABLED`.