MCPClient: single-threaded implementation
This commits implements a simpler version of MCPClient that runs in a single
thread and just running one worker. It installs a signal handler to make sure
that the process exits when `SIGTERM` is received.

It also dies when Gearman server is available. Instead of implementing the
retry ourselvs we rely on the container scheduler to start a new instance for
us.