Fix Handle config form bug
Before rendering the Handle config form in the dashboard, the view used to
expect a ``'pid_request_verify_certs'`` key in ``settings_dict``.
However, this key will not be present if the form has never been saved
and a ``KeyError`` will be raised. This change sets the default for this
key to ``True``.