MCPClient: add single-threaded client
This commit adds a new executable `archivematicaSingleClient.py` that can be
used to execute a single-threaded MCPClient worker. This commit also updates
the Dockerfile so the new ENTRYPOINT points to this executable, this is because
Docker container are frequently executed by a scheduler where it is easy to
manage multiple replicas.