Fix ASpace DIP upload, refs #746
The ASpace config dict is used in linkTaskManagerReplacementDicFromChoice
to set the parameters sent to the upload-archivesspace script, which is
expecting restrictions instead of premis. This field doesn't seem to
be used anywhere else, so it only needs to be updated in the related form
fields that are used to update the config dict.

This also fixes an issue where the config port was not being used in the upload.