storageService: add new query timeout
This is in addition to the existing client timeout that already exists in the
storageService module. This new timeout is used in requests that we know they
should be processed immediately.

Implementation detail: `_storage_api_session` is now used in requests that are
supposed to be processed quickly (including async). A new private function
called `_storage_api_slow_session` is used in those requests that still take
too long which is undesirable but we haven't had the time to fix yet.

This is connected to #1133.