Dashboard: update /api/transfer/start_transfer
In a previous commit we have added to MCPServer the ability to create packages
and start tranfers. This commit updates `/api/transfer/start_transfer/` and the
underlying views to make use of this new functionality. In particular, it uses
the `auto_approve=False` mode to maintain backward compatibility - this method
also blocks until the files are copied/moved to the activeTransfer directory.