MCPClient: fix matching originals to manual derivatives
Fixes the matching of original files to their manually created
preservation derivatives in a manualNormalization/preservation/
directory. Previous behaviour would find a match and all other Files
that matched by having a path that was a superfix of the original path.
This would trigger an exception. This fixes the issue by returning the shortest
match. It also adds print statements to better understand what has happened
through the Tasks GUI.