Explicitly handle Archivematica Archivesspace cfg
Because we are able to identify in our own configuration that the host
string to conect to Archivesspace is still blank we know that it hasn't
been configured. In those instances, we can raise a custom exception
stating as much, instead of attempting to create an Archivesspace client
and handling an InvalidURL exception.

Resolves #967