Change filegrpuse of SIP pointer files
In AM 1.6, the pointer file was being created in the MCPClient and added
to the SIP files with the default filegrpuse: 'original'. This was making
this file to appear, named 'xml', in the normalization reports as an
error. This file should have been added to the SIP files with
'submissionDocumentation' as its filegrpuse and changing it in this
migration will make them disappear from existing reports.