MCPClient: job email report follow-up (#69)
* CR: pep8 79 cols

For new code we try to follow the conventions established in PEP 8. This commit
limit all lines added or edited to a maximum of 79 characters.

* CR: stop using emailOnCompletion as entry point

We can add the code back if we decide in the future that this module should
become a client script. In the meanwhile, it can be considered dead code and
the practice is to avoid dead code as much as possible.

* CR: move emailOnCompletion to lib

We created the lib package for cases like this. It was empty until know but
the idea is to start moving things here as they're added to the project.

* CR: rename module emailOnCompletion » job_email_report

I think that the new name describes better what the module does instead of
describing how it was used by its consumer.

* CR: remove stdout parameter

There is no use case for this. If you want to verify that the email works just
use the console email backend which is enabled by default in the development
environment or in production when no email settings are provided.

* CR: stop run_job from returnig exit code

It was unused. Also, it's more Pythonic to raise an exception instead which
will effectively interrupt the execution of the client script and have it
return the code `1`.

* CR: avoid bare except

Linters don't like plus it's better to determine if you want `BaseException` or
`Exception` instead.

* CR: add module-level docstring

* CR: fix docstrings, use imperative mood

* CR: s/attachmentText/attachment_text

* CR: i18n fixes

- I18nize missing strings
- Avoid concatenation, splitting sentences into several keys presumes a
  grammar and a certain sentence structure.

* CR: clean up constants

* CR: delete unneeded django.setup() call

Now that this module is not an entry point we don't need this.

* CR: stop using TEST_MAIL

The console email backend should be enough. Send email to users just check the
output with `docker-compose logs -f archivematica-mcp-client` instead of
excepting that the message is going to be delivered.

* CR: s/run_job/run_report

I personally think that this is a better name for the function.

* CR: post_store_aip_hook to handle exception

* CR: add missing ending period

* CR: the list of jobs can be empty

* Fix import error I caused in prev commit

* Added remaining standard messages to the migration

* Adjusted formatting