MCPClient: OCR transcrs are not derivs on reingest
Stops archivematicaUpdateSizeAndChecksum.py from treating OCR
transcription files as derivatives on full re-ingest. Because OCR files
are deleted on full re-ingest, the previous behaviour was causing a
failure when OCR transcriptions were documented in the METS file but
preservation derivatives were not. The presence of preservation
derivatives (which are not deleted form the Files table) was masking this issue
in typical transfers. See #1043.

Fixes #1043