Reingest: fix AttributeError MDRef
metsrw assumes that the METS is self-contained, MDRef pointers are not
followed. It would be up to `archivematicaCreateMETSReingest.py` to do so but
instead this commit ignores them. We're doing this because any rights updates
performed during re-ingest would result in the AIP METS file being edited, not
the DSpace one.

This fixes #1083.