Make fixes to CONTRIBUTING.md
This commit adds references to GitHub permalinks which may provide a
more accurate historical picture of the various Artefactual
repositories over a period of time should they be used effectively by
contributors.

We have also updated the process guidance for logging issues, as well
as elaborating on the style to adopt when creating commit messages.
References have been added to Pylint as a potential tool to
aid in code-review. Markdown style is now more consistent with the
Archivematica docs style-guidance.

* Resolves #1087
* Resolves #876
* Resolves #1088
* Resolves #860