Update defaultNextChainLink for extract packages
There are two extract packages microservices which continue to process
data even if they have failed to extract content, or the service
itself has failed for some other reason. Make sure that the Task falls
back to 'Failed Transfer' so that the user knows and a decision can be
made based on the content of the transfer.