Add site_url setting
In this commit the Dashboard learns a new configuration attribute named
`site_url`. This allows the Dashboard to share with Storage Service a URL that
is accessible over the network. Before this commit, Storage Service could only
tell the remote address during the pipeline registration process by looking up
the value in the `REMOTE_ADDR` header which is not always sufficient.

We're giving users the possibility to define the value of `site_url` via
config (INI or environment) or via the user interface (installer or admin page)
which is persisted using the `DashboardSetting` model. When the value is given
via config (INI or environment), the user interface form field is shown in read
only mode (or disabled).

The `storageService` module is updated so its `create_pipeline` function takes
the `remote_name` given by the user. This attribute is already accepted by the
Storage Service API.