Docker: use Debian jessie base images
Our images were based on `python:2.7`. This image moved to Debian stretch
causing an issue in our build. For further details see https://git.io/fCV6c.

This commit moves the base images to `python:2.7-jessie` to ensure that we
continue using Debian jessie until we have time to evaluate the upgrade.

This is connected to #1124.