Correct PREMIS:originalName for packaged objects
This commit does a number of things. First, the logic for extracting
zipped packages is reversed so the package is temporarily renamed so
they can be extracted into a folder that reflects their original
name.

Second the commit adds behavior that allows the originalLocation to
be set verbosely inside the code.

This seems to be the least invasive way to change the Archivematica
code to be able to generate directory uuids and still pair that
information with PREMIS information where needed.

Unit tests have been added to show the impact of adding original
location.

Resolves #1094.