Add retry logic for concurrency sensitive database transactions
When the MCP Server inserts rows into the `Tasks` table, it holds a
lock on the table primary key index.  When the MCP Client updates
existing tasks to set their start time, it does the same.  If these
transactions overlap, MySQL may detect this condition as a deadlock
and abort one the transactions with a retriable/retryable(?!)
exception.

To handle this, we can retry the failed transaction after a short
wait.  It should always succeed eventually.

Addresses issue #1198