MCPClient: use auto_close_db
Make sure that db connections created by the Gearman handler don't remain
open. We do this to avoid reusing connections that are unusable because the
server closed them due to inactivity.

The problem that this commit is trying to solve is easily reproducible when
`wait_timeout` in `mysqld` has a low value, e.g. 10 seconds. The default
is much higher (8 hours) so the problem only appears after the system
has been unused for a long period of time.

This is connected to https://github.com/archivematica/Issues/issues/52.