Dashboard: retrieve config using UUIDs
This commit updates our `ingest.js` and `jobs.js` files so they don't
use labels for conditional behaviour. Instead, we're using UUIDs which
are going to be the same regardless the language used in the user
interface.

Also, this commit is removing some dead code:
- Review button in Store DIP (broken and unused).
- Dialog in unexistent micro-service:
  "Normalize submission documentation to preservation format".

This is connected to https://github.com/artefactual/archivematica/issues/1101.
This is connected to https://github.com/archivematica/Issues/issues/62.