Create a Dataverse transfer type
This commit Creates a Dataverse transfer type. It consists of a
migration that places two Dataverse specific microservices into
the Archivematica workflow for when a Dataverse transfer type is
selected.

The microservices introduced are:

* convert_dataverse_structure.py
* parse_dataverse_mets.py

Convert Dataverse Strcture creates a Dataverse specific METS file
describing the contents of a Dataverse transfer. This work was
previously part of the Automation Tools pre-transfer functionality
but it made sense to bring the workflow inside Archivematica and
remove that extraneous step.

Parse Dataverse METS validates the contents of the Dataverse METS
and commits its contents to the database for the creation of the
SIP prior to ingest.

AIP METS generation has been enhanced to know how to manipulate the
data created during a Dataverse transfer.

* Resolves #1147
* Resolves #1163
* Resolves #1149
* Resolves #1121
* Resolves #1119
* Resolves #1148
* Resolves archivematica/Issues#57
* Resolves archivematica/Issues#59
* Resolves archivematica/Issues#104
* Resolves archivematica/Issues#131
* Resolves archivematica/Issues#152
* Resolves archivematica/Issues#187