Populate processing config fields in MCPServer
Similar to #1238, we want MCPServer to be the only component interacting
with workflow models. In this commit, the processing configuration form
is updated so all the workflow-related data is pulled from MCPServer via
the RPC interface.

To support this, MCPServer learns a new RPC `getProcessingConfigFields`
and a new `processing_config` module.

This is connected to #1101.