MCPClient: force inline execution of validate_file
This commit ensures that we only have one instance of MediaConch at a
time to avoid the problem described in the issue linked below. This is
achieved by removing the `concurrent_instances` attribute from the
client script which forces the code in it to be executed inline by
MCPClient.

Connects to https://github.com/archivematica/Issues/issues/44.