Dashboard: fix glitch with long names
The file browser wasn't handling text overflow properly - the expand
button was being shadowed by the directory name. This commit changes the
layout of directory items so it behaves as a grid where the button and
the text behave as columns within its row.

This connects to https://github.com/archivematica/Issues/issues/53.