Don't return files removed from transfer
When extract contents runs on a Dataverse transfer it has the
potential to remove Dataverse generated files from the transfer where
those files are packaged as a zip, e.g. RData. In this instance we
preserve the content not the container.

If the file has been purposely removed we can safely let the
microservice continue processing.

Logging has been reduced to make it more helpful in the primary
function associated with this fix.

Tests have been added to help refactor and maintain the function
going forward.