Don't return files removed from transfer
When extract contents runs on a Dataverse transfer it has the
potential to remove Dataverse generated files from the transfer
where those files are packaged as a zip, e.g. RData. The contents
of the RData package remains, but the original container doesn't.

With this fix we no longer try to work with the data where it has
been removed from the transfer.