Enable use of identifiers.json for PIDs in PREMIS
Where an identifiers.json file is found in a transfer metadata folder
and Bind PIDs is configured to 'Yes'; parse the identifiers.json file
and record the discovered PIDs in PREMIS.

If the Bind PIDs is configured to Yes and the handle server
configuration is missing or incorrect then the identifiers.json will
still be used.

Identifiers can be recorded for the SIP, directories, and files.

Some redundancy has been removed between Bind PID modules though some
may remain.

New steps have been taken to make sure services exit more gracefully
if there are issues around configuration, e.g. if the required fields
have not been populated for the handle server, the the suggestion is
that there is zero configuration. This should not result seemingly in
a failure for day-to-day users.

Flake8 and Pylint changes have been included in this commit to help
improve the robustness of the code.