Add JSON-encoded workflow
0) Add workflow and schema

The schema was written manually.

The workflow as generated via `json_links` using the `--locale-dir`
option in order to extract the English messages and pull the latest
translations so they're also included in the document.

1) Add workflow module

This commit adds a new `workflow` module that is capable to read and
validate the JSON-encoded workflow and provide a minimal interface so it
can be browsed by MCPServer.

The workflow document is loaded by `archivematicaMCP.py` (our entry point).
Then it is passed to the different subsystems, i.e. directory watcher
(in this commit) and `jobChain` instances (in later commits).

2) Use new workflow in RPCServer

RPCServer becomes a subclass of `GearmanWorker` so it's easier to have
the workflow data shared across all the handlers.

The `abilities` module is now moved to MCPServer since it's not shared
with the Dashboard anymore. This is because Dashboard is becoming a RPC
client, business logic is now in MCPServer. Also, added some tests.

3) Use new workflow in MCPServer engine

4) Delete workflow models

This commit deletes the workflow-related models in the `main` app,
replaced by the JSON-encoded workflow.

- MicroServiceChain
- MicroServiceChainChoice
- MicroServiceChainLink
- MicroServiceChainLinkExitCode
- MicroServiceChoiceReplacementDic
- StandardTaskConfig
- TaskConfig
- TaskConfigAssignMagicLink
- TaskConfigSetUnitVariable
- TaskConfigUnitVariableLinkPull
- TaskType
- WatchedDirectory
- WatchedDirectoryExpectedType

Two `ForeignKey` fields had to be migrated to `UUIDField`:

- job.microservicechainlink
- unitvariable.microservicechainlink

5) Update Dashboard

These are some changes needed in the Dashboard which performs as a
RPCServer client.

- Pull status of units from the new RPC.
- Send user_id (uid) and language to server.
- Tweaks in JavaScript code to show translations.
- Cleanup: get_directory_name becomes a method of `Job`.
- Cleanup: delete dead code around subjobs.

6) Update Dashboard - microservice list

Add new `getUnitStatus` RPC that returns details of individual units.
I've used this new method in the transfer/ingest microservice list page.
The new method is different in that MCPServer returns the details
translated to the language requested as it has access to the workflow
dataset.

The `Job` model learns how to return the directory name replacing the
old helper under `contrib`.

7) Delete json_links and graph_links

These are based on database models which have been removed. We're
building a new tool to achieve the same gaols in addition to be able to
alter the workflow encoded in JSON.