Extract multiple Dataverse authors
This commit will see the extraction of multiple author entities from
the Dataverse metadata dataset.json file. Affiliation is also added
to the DDI segment. There is a small refactor to separate some of the
responsibilities of the code to better enable future refactoring.
Tests have been added to test the addition of multiple authors.