Avoid SS locations overwrite
Thanks to Ross Spencer:

Previously AIP or DIP storage locations without labels would overwrite
each other in the dict created to process a user choice. Here we
change that structure to something more robust so that this doesn't
happen and the user can see all of their locations with some indicator
of what that location is whether or not it has a description in the
storage service.

Previously an AIP or DIP storage location in the storage service
without a description would present as a blank field in the
administration form. This is corrected to show the relative_path
when a description hasn't been specified. Similar to the behavior
in the ingest workflow.

Squashed commit of the following:

commit 2a115969e575e70aecac1b72d4b006eaef4d6b9d
Author: José Raddaoui Marín <raddaouimarin@gmail.com>
Date:   Tue Feb 12 20:13:33 2019 +0100

    Cosmetic changes

commit 185e8dd519496e120b9979306727fb99bcc8a3fe
Author: José Raddaoui Marín <raddaouimarin@gmail.com>
Date:   Tue Feb 12 20:12:34 2019 +0100

    Fix logging

commit 125d007cf63dea5b02eefd55a7b3121d21464d67
Author: Ross Spencer <all.along.the.watchtower2001@gmail.com>
Date:   Fri Feb 8 23:20:41 2019 +0100

    Provide users with more AIP/DIP store description

    Previously an AIP or DIP storage location in the storage service
    without a description would present as a blank field in the
    administration form. This is corrected to show the relative_path
    when a description hasn't been specified. Similar to the behavior
    in the ingest workflow.

commit 3ba964a6fd1ce9d9402164e9ddd380343b99ec27
Author: Ross Spencer <all.along.the.watchtower2001@gmail.com>
Date:   Fri Feb 8 23:10:06 2019 +0100

    Use a JSON structure to pass AIP/DIP store data

    Previously AIP or DIP storage locations without labels would overwrite
    each other in the dict created to process a user choice. Here we
    change that structure to something more robust so that this doesn't
    happen and the user can see all of their locations with some indicator
    of what that location is whether or not it has a description in the
    storage service.