Delete workflow content types
After the deletion of workflow models we need to delete the corresponding
content types. It's probably not harmful to leave them behind but Django is
asking users to delete staled content types when `migrate` is used which may
be confusing.

This is connected to https://github.com/archivematica/Issues/issues/514.