RPCServer: restore access_system_id property
The status of ingest/transfer was moved to MCPServer and merged into a single
function. A branch in the code responsible for populating `access_system_id`
for SIP objects was not being executed because a condition was not adequate.
The behaviour expected in the frontend wasn't triggered because
`access_system_id` was not being included in the status data. This commit
restores the property.

Introduced by https://github.com/artefactual/archivematica/pull/1251.

This is connected to https://github.com/archivematica/Issues/issues/528.