MCPClient: set activeAgent fallback in ingest
Set `activeAgent` to the value used in Transfer to ensure the existence of the
user agent in PREMIS events. This value may change as soon as a different user
interacts with the user interface in processing, but this commit is useful when
interactions do not occur.

Follow-up for https://github.com/artefactual/archivematica/pull/1354.

Connects to https://github.com/archivematica/Issues/issues/529.