Bag transfers before sending to backlog
Transfer indexing is now done from `move_to_backlog` - this removes some noise
from the user interface while maintaining the same functionality.

`elasticSearchFunctions` learned how to look up files in the database during
indexing of `transferfiles` documents when the transfer package is a BagIt -
this may not always occur, e.g. when backlog contains transfers created by
older releases of the system.

`filesystem_ajax` learned how to arrange a SIP when the source is a transfer in
BagIt format. Compatibility with non-baggited transfers is maintained. Names
of private functions in `filesystem_ajax.views` have been prefixed with an
undercore.