Ensure file paths aren't turned into dir paths
Previously when transfers of individual files - possible for DSpace
and Zipped Bag transfers were sent to Archivematica they were
incorrectly turned into paths that would be interpreted as
directories preventing approval of a transfer for not being able to
locate that directory. Here we re-enable individual DSpace file
transfers and galvanize the check to also test whether or not we're
looking at a file at the same time.