Introduce PID declaration feature
Developed for the International Institute of Social History (IISH) in
The Netherlands. Archivematica users can supply custom identifiers to
an AIP and AIP contents by making use of an identifiers.json file in
the metadata folder of the transfer. In this commit we enable this
feature. We add corresponding unit tests to ensure that PID
declaration and Bind PID features work alongside each other as well
as alongside Archivematica's own UUID for premis:Objects.

Fixtures and tests are also added to the Bind PIDs feature as it is
currently understood in Archivematica.

Parse Dataverse tests have been updated to be more specific in what
they are looking to retrieve from the database.

We have also added a little more feedback to the user about why Bind
PIDs hasn't worked if it is toggled on but hasn't been configured
correctly.