Arrangement: avoid local access to determine BagIt format
When a SIP is arranged, we need to ask SS which files are copied to the
local staging path.

`logs` and `metadata` are also copied because they're used in Ingest to
build the final SIP. However these entries are not listed in `SIPArrange`
so we need to guess their paths which is going to be different depending
on the format of the transfer package, e.g. `data/logs` instead of `logs`
when it's using the BagIt format.

This commit changes the logic that guesses the format of the package.
Originally, we assumed that physical access to the transfer package was
possible to determine its format. As that's not always true, we're now
guessing the format by looking at existing entries listed by `SIPArrange`.