Separate validation responsibilities
We want to validate the handle server configuration early, but only
on the values we know that we will have. This commit sees a change to
the Bind PIDs code that pulls apart fields required to bind and
attach PIDs in Archivematica but calls their validation methods at
more appropriate points in the execution of the code.