Add a METS download button
We introduce a METS download button to the Archival Storage Tab.

Users can now use a button to stream AIP METS directly from a
package for review. METS can be consumed for a variety of reasons
and so making it easier to review without downloading an entire
archival information package (AIP) has its benefits.

The feature introduces the AMClient (Archivematica API Client)
library into Archivematica core which provides helpers for accessing
Archivematica API functionality. This is in-line with the recent
Architectural Decision Record: ADR-0009.

Smaller changes are to the test layout used by the Archival Storage
Tab component, as well as the introduction of tests for the new
capability and previous file-streaming capability to aid in the
refactor of these functions in the future.