Provide better handling of custom structMap errs
A number of different error conditions were spotted during testing of
the Archivematica 1.10 release candidate. Rather than failing
silently in the system we introduce logic to make the output more
robust. Example failure conditions for custom structMaps include when
files do not exist, or there are empty attributes for the CONTENTID
field.