Initial package scheduler
A starting point for making MCPServer more package aware.

First pass at a package scheduler. The concept is we keep queues at
both the package and job levels, and only move on to a new package
when one has been completed or failed.

Also:
    * renames the old concept of unit to package and refactor
    * switch to new style classes (!)
    * refactor job chain logic
    * refactor link task manager classes
    * improves efficiency by not walking the entire filesystem on every
      linkTaskManagerFiles run
    * consolidate XML processing logic
    * submit gearman tasks (and check status) from each executor thread,
      run by the package scheduler.

Note: **this change removes support for delayed task execution**.