Initial package scheduler
First pass at a package scheduler. The concept is we keep queues at
both the package and job levels, and only move on to a new package
when one has been completed or failed.

This seems to work pretty well in practice, although ofter a waiting
transfer will be started between processing a transfer and its SIP -
this seems to be because there is a delay in queuing the SIP job chain.