Remove mutable default argument
When reindexing, `_list_files_in_dir()` is run once for each item. If
the `filepaths` argument is left set to `[]`, then each item will
erroneously include all previously indexed files.

By setting it to `None` and initing the `[]` default inside the function
body, we avoid this undesirable behavior because the list is inited anew
each time the function runs.