Fix constant gearman polling in MCPServer
An overly tight loop in GearmanTaskBackend resulted in constant polling
of gearman for status updates while waiting for a job to complete,
which burns a lot of CPU for no reason. Now we use a lower level
client method to wait for job activity.

Also removes possible naming confusion by renaming `gearman` to
`gearman_backend`, so it doesn't have the same name as the library.