elasticsearch: parse size in index_aip_and_files
The aip_size in index_aip_and_and_files is assumed to be an int. This
was the case until we changed the size field in the Storage Service
Package form Integer to BigInterger. Because of this change requests
deserializes the size field as an string, instead of an int.

To make sure that such arbitrary input changes cannot break this
function we convert the aip_size to an int. This means that if the input
is of type int, we convert the int to an int. If the input type is a
string, we convert the string to an int.